+++ src/java/abfab3d/grid/RowOfInt.java
@@ -18,7 +18,7 @@
 public interface RowOfInt {
     
     public int get(int x);
-    public void set(int x, int value);
+    public void set(int x, long value);
     public Object clone();
     public void clear();   // clear data, but keep the memory 
     public void release(); // release all the memory
